app/requirements.txt
attrs==21.4.0
autopep8==1.6.0
click==8.1.3
coverage==6.4.2
dnspython==2.2.1
Flask==2.1.3
Flask-Cors==3.0.10
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pep8==1.7.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
Werkzeug==2.1.2
